Update patch set 13
Patch Set 13:

FYI, what's exactly the point of not submitting the client to Fedora?
Is it a process issue? (As a sponsor, I can help with that)

For the record, we want to have clients to be available as widely as possible. We can still do the review in gerrit but in the end, formal review will be posted on bugzilla for both RDO and Fedora reviews, so little difference.

Patch-set: 13