Update patch set 13
Patch Set 13:

Hi Javier and hguemar,

We would love to make karborclient accessable on both RDO and Fedora.
In fact, I totally have no idea of packaging for Fedora since I'm not familiar with it. Could that be done in this patch or is it just another kind of process similar to RDO packaging? If it's another process, then I may need your assistance.

Patch-set: 13